# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/class": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "강의개설",
                "parameters": [
                    {
                        "description": "강의정보",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Class"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "강의 생성에 실패 했습니다.",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/class/enroll/{classCode}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "수강신청",
                "parameters": [
                    {
                        "type": "string",
                        "description": "학수번호",
                        "name": "classCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "학수번호가 유효하지 않습니다.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "강의가 존재하지 않습니다..",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/quizsets/quizset/{quizSetId}/class/{classCode}": {
            "post": {
                "description": "퀴즈 셋 고유 아이디와 학수번호를 받아서 클래스의 퀴즈셋 목록에 추가한다.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quiz set"
                ],
                "summary": "유저의 퀴즈셋에 있는 퀴즈를 강의의 퀴즈셋으로 가져온다",
                "parameters": [
                    {
                        "type": "string",
                        "description": "클래스에 불러오려는 퀴즈셋의 아이디",
                        "name": "quizSetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "퀴즈를 추가하려는 강의의 학수번호",
                        "name": "classCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "CLASS_QUIZ_SET_ALREADY_EXISTS",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "퀴즈 셋 고유 아이디와 학수번호를 이용해서 클래스에서 퀴즈셋을 삭제한다.\n원본 퀴즈 셋은 삭제되지 않는다.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quiz set"
                ],
                "summary": "클래스에 있는 퀴즈셋을 삭제한다",
                "parameters": [
                    {
                        "type": "string",
                        "description": "강의 에서 삭제 하려는 퀴즈셋의 아이디",
                        "name": "quizSetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "강의의 학수번호",
                        "name": "classCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "CLASS_QUIZ_SET_NOT_EXISTS",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/quizsets/quizset/{quizSetId}/quiz/{quizId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quiz set"
                ],
                "summary": "퀴즈셋에 있는 퀴즈를 삭제한다",
                "parameters": [
                    {
                        "type": "string",
                        "description": "퀴즈를 삭제 하려는 퀴즈셋의 아이디",
                        "name": "quizSetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "삭제 하려는 퀴즈의 아이디",
                        "name": "quizId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "요청이 올바르지 않습니다.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/quizsets/quizset/{quizsetId}/quiz": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quiz set"
                ],
                "summary": "퀴즈셋에 퀴즈를 추가한다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "퀴즈를 추가하려는 퀴즈셋의 아이디",
                        "name": "quizSetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "퀴즈 정보",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.QuizCreateForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "요청이 올바르지 않습니다.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/quizsets/users/{username}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quiz set"
                ],
                "summary": "유저가 만든 모든 퀴즈셋 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "유저 로그인 아이디",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "퀴즈셋목록(배열)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.QuizSetGetForm"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/classes/managing/{userName}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "유저가 관리하고 있는 모든 강의들을 가져온다",
                "parameters": [
                    {
                        "type": "string",
                        "description": "유저 아이디",
                        "name": "userName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "강의목록(배열)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Class"
                            }
                        }
                    },
                    "400": {
                        "description": "학수번호가 유효하지 않습니다.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "강의가 존재하지 않습니다..",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Class": {
            "type": "object",
            "properties": {
                "class_code": {
                    "type": "string",
                    "example": "형식언어"
                },
                "class_name": {
                    "type": "string",
                    "example": "CSE-1234"
                }
            }
        },
...
        "dto.QuizCreateForm": {
            "type": "object",
            "properties": {
                "quiz_answer": {
                    "type": "string",
                    "example": "퀴즈 정답"
                },
                "quiz_content": {
                    "type": "string",
                    "example": "퀴즈 렌더링 정보"
                },
                "quiz_title": {
                    "type": "string",
                    "example": "퀴즈 문제"
                },
                "quiz_type": {
                    "type": "string",
                    "example": "퀴즈 타입(객관식, 주관식)"
                }
            }
        },
        "dto.QuizGetForm": {
            "type": "object",
            "properties": {
                "quiz_answer": {
                    "type": "string",
                    "example": "퀴즈 정답"
                },
                "quiz_content": {
                    "type": "string",
                    "example": "퀴즈 렌더링 정보"
                },
                "quiz_id": {
                    "type": "integer",
                    "example": 1234
                },
                "quiz_title": {
                    "type": "string",
                    "example": "퀴즈 문제"
                },
                "quiz_type": {
                    "type": "string",
                    "example": "퀴즈 타입(객관식, 주관식)"
                }
            }
        },
...
        "dto.QuizSetGetForm": {
            "type": "object",
            "properties": {
                "quiz_set_author_name": {
                    "type": "string",
                    "example": "퀴즈셋 생성자 로그인 아이디"
                },
                "quiz_set_id": {
                    "type": "integer",
                    "example": 1234
                },
                "quiz_set_name": {
                    "type": "string",
                    "example": "퀴즈셋 이름"
                },
                "quizzes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuizGetForm"
                    }
                },
                "total_score": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
    }
}